Clean up `ci` test
By using `required-features`, one now gets an error from `cargo` when
one tries to run `cargo test --test ci` without `--features=ci`.

And one still gets a warning that the feature is disabled when one runs
just `cargo test`.
