--- core/tests/ci_is/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(not(feature = "ci"))]
-mod disabled;
-
-#[cfg(feature = "ci")]
-mod enabled;
