+++ core/tests/ci_is_disabled.rs
@@ -1,7 +1,9 @@
+#![cfg(not(feature = "ci"))]
+
 use std::io::{stderr, Write};
 
 #[test]
-fn ci_is_disabled() {
+fn warn() {
     #[allow(clippy::explicit_write)]
     writeln!(stderr(), "Warning: feature `ci` is disabled").unwrap();
 }
